[FIX] don't recompute analytic lines for draft move in analytic
distribution wizard

+++ bankayma_account/wizards/account_move_update_analytic.py
@@ -72,7 +72,8 @@ class AccountMoveUpdateAnalytic(models.TransientModel):
                     "line_ids": changed_move_lines,
                 }
             )
-            move.line_ids.analytic_line_ids.unlink()
-            move.line_ids._create_analytic_lines()
+            if move.state == "posted":
+                move.line_ids.analytic_line_ids.unlink()
+                move.line_ids._create_analytic_lines()
 
         return result
